Move Size<T> to the Implementation namespace
