+++ src/Count.cs
@@ -1,7 +1,7 @@
 namespace Fuzzy
 {
     /// <summary>Specifies an inclusive range for the number of elements in a fuzzy collection.</summary>
-    public sealed class Count : Size<Count>
+    public sealed class Count : Implementation.Size<Count>
     {
     }
 }
