+++ src/Length.cs
@@ -1,7 +1,7 @@
 namespace Fuzzy
 {
     /// <summary>Specifies an inclusive range for the length of a fuzzy string or array.</summary>
-    public sealed class Length : Size<Length>
+    public sealed class Length : Implementation.Size<Length>
     {
     }
 }
