+++ test/LengthTest.cs
@@ -2,6 +2,6 @@ namespace Fuzzy
 {
     public class LengthTest
     {
-        readonly Size<Length> sut = new Length();
+        readonly Implementation.Size<Length> sut = new Length();
     }
 }
